RDCC-630/718 - created a validator class and moved validator methods of PBA to the new class. Created a REGEX to combine the two we currently have
