+++ src/main/java/uk/gov/hmcts/reform/professionalapi/service/PaymentAccountService.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.reform.professionalapi.service;
 
-import java.util.Set;
-
 import uk.gov.hmcts.reform.professionalapi.controller.request.PbaEditRequest;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 import uk.gov.hmcts.reform.professionalapi.domain.PbaResponse;
@@ -17,6 +15,4 @@ public interface PaymentAccountService {
     void addPaymentAccountsToOrganisation(PbaEditRequest pbaEditRequest, Organisation organisation);
 
     PbaResponse addUserAndPaymentAccountsToUserAccountMap(Organisation organisation);
-
-    void validatePaymentAccounts(Set<String> paymentAccounts, String orgId);
 }
