+++ src/main/java/uk/gov/hmcts/reform/professionalapi/service/impl/PaymentAccountServiceImpl.java
@@ -2,21 +2,18 @@ package uk.gov.hmcts.reform.professionalapi.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.reform.professionalapi.configuration.ApplicationConfiguration;
 import uk.gov.hmcts.reform.professionalapi.controller.feign.UserProfileFeignClient;
-import uk.gov.hmcts.reform.professionalapi.controller.request.InvalidRequest;
 import uk.gov.hmcts.reform.professionalapi.controller.request.PbaEditRequest;
 import uk.gov.hmcts.reform.professionalapi.domain.*;
 import uk.gov.hmcts.reform.professionalapi.persistence.PaymentAccountRepository;
@@ -121,43 +118,8 @@ public class PaymentAccountServiceImpl implements PaymentAccountService {
 
     public List<UserAccountMapId> generateListOfAccountsToDelete(ProfessionalUser user, List<PaymentAccount> accounts) {
 
-        List<UserAccountMapId> userAccountMapIds = accounts.stream().filter(account -> null != user && null != account)
+        return accounts.stream().filter(account -> null != user && null != account)
                 .map(account -> new UserAccountMapId(user, account))
                 .collect(Collectors.toList());
-
-        return userAccountMapIds;
     }
-
-    public void validatePaymentAccounts(Set<String> paymentAccounts, String orgId) {
-        if (!paymentAccounts.isEmpty()) {
-            checkPbaNumberIsValid(paymentAccounts);
-            checkPbasAreUniqueWithOrgId(paymentAccounts, orgId);
-        }
-    }
-
-    public void checkPbaNumberIsValid(Set<String> paymentAccounts) {
-        String invalidPbas = paymentAccounts.stream()
-                .filter(pbaAccount -> pbaAccount == null || !pbaAccount.matches("(PBA|pba).*") || !pbaAccount.matches("^[a-zA-Z0-9]+$"))
-                .collect(Collectors.joining(", "));
-
-        if (!StringUtils.isEmpty(invalidPbas)) {
-            throw new InvalidRequest("PBA numbers must start with PBA/pba and be followed by 7 alphanumeric characters. The following PBAs entered are invalid: " + invalidPbas);
-        }
-    }
-
-    public void checkPbasAreUniqueWithOrgId(Set<String> paymentAccounts, String orgId) {
-        List<PaymentAccount> paymentAccountsInDatabase = paymentAccountRepository.findByPbaNumberIn(paymentAccounts);
-        List<String> uniquePBas = new ArrayList<>();
-
-        paymentAccountsInDatabase.forEach(pbaInDb -> paymentAccounts.forEach(pba -> {
-            if (pbaInDb.getPbaNumber().equals(pba) && !pbaInDb.getOrganisation().getOrganisationIdentifier().equals(orgId)) {
-                uniquePBas.add(pba);
-            }
-        }));
-
-        if (!uniquePBas.isEmpty()) {
-            throw new InvalidRequest("The PBA numbers you have entered: " + String.join(", ", uniquePBas) + " belongs to another Organisation");
-        }
-    }
-
 }
